Use DTC from prebuilts
Sibling CL I2444a8574cf41e0ca92cf92b9bd4aa3a2de8b4ec adds a prebuilt
of the Device Tree Compiler. Switch our scripts to using it instead of
the one in system.

This is necessary because:
(a) hermetic builds, and
(b) dtc in Ubuntu is outdated.

Note that we cannot remove dtc from the compilation container yet as it
is still required for compiling the Linux kernel.

Bug: 117551352
Bug: 132428451
Change-Id: I68a9c266b0212d8bbfb60c5389b28d3f6fe5e8cf